Fixed issues with the payday method of the movie class
Fixed the payday method to accept an ArrayList (instead of vector) of the classes required and added the pay earned from each object in the lists to the total amount spent. Also got rid of redundant methods in the PA class and fixed the Person abstract class to better represent the repeated elements.